requirements.txt
astroid==2.2.5
bcrypt==3.1.7
beautifulsoup4==4.8.1
certifi==2019.6.16
cffi==1.12.3
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.1
django-bootstrap4==1.1.1
django-model-utils==3.1.2
django-tastypie==0.14.2
djangorestframework==3.8.2
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pipenv==2018.11.26
psycopg2==2.8.4
pycparser==2.19
pylint==2.3.1
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.1
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.8
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
whitenoise==5.0.1
wrapt==1.11.2
